refactor: add main and app.module to infra layer
