+++ src/infra/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { PrismaModule } from '@/infra/database/prisma.module';
-import { AuthModule } from './infra/auth/auth.module';
-import { HttpModule } from './infra/http/http.module';
+import { AuthModule } from '../infra/auth/auth.module';
+import { HttpModule } from '../infra/http/http.module';
 
 @Module({
   imports: [
